+++ packages/fxa-content-server/app/tests/spec/views/mixins/form-prefill-mixin.js
@@ -7,15 +7,19 @@ define(function (require, exports, module) {
 
   const { assert } = require('chai');
   const Backbone = require('backbone');
-  const FormView = require('views/form');
   const Cocktail = require('cocktail');
+  const FlowEventsMixin = require('views/mixins/flow-events-mixin');
+  const FormView = require('views/form');
   const FormPrefillMixin = require('views/mixins/form-prefill-mixin');
+  const Notifier = require('lib/channels/notifier');
+  const sinon = require('sinon');
 
   const View = FormView.extend({
     template: () => `
       <input class="nameless" />
       <input id="id-only" />
       <input name="already-filled" value="this is filled in" />
+      <input name="empty" />
       <input name="has-data-novalue" data-novalue />
       <input name="name-only" />
       <input name="name-preferred" id="id-ignored" />
@@ -23,19 +27,23 @@ define(function (require, exports, module) {
       <textarea name="textarea" />
     `
   });
+
   Cocktail.mixin(
     View,
+    FlowEventsMixin,
     FormPrefillMixin
   );
 
   describe('views/mixins/form-prefill-mixin', () => {
     let formPrefill;
+    let notifier;
     let view;
 
     beforeEach(() => {
       formPrefill = new Backbone.Model({
         'id-only': 'id only value',
         'already-filled': 'a different already-filled value', //eslint-disable-line
+        'empty': '',
         'has-data-novalue': 'has-data-novalue value',
         'name-only': 'name only value',
         'name-preferred': 'name preferred',
@@ -43,22 +51,36 @@ define(function (require, exports, module) {
         'textarea': 'the value for the text area'
       });
 
+      notifier = new Notifier();
+
       view = new View({
-        formPrefill
+        formPrefill,
+        notifier
       });
 
+      sinon.spy(view, '_engageFlowEventsForm');
+
       return view.render();
+    });
+
+    afterEach(() => {
+      view.destroy(true);
+      view = null;
     });
 
     it('pre-fills input elements w/ name/id without `autocomplete=off` attribute', () => {
       assert.equal(view.$('.nameless').val(), '');
       assert.equal(view.$('#id-only').val(), 'id only value');
       assert.equal(view.$('[name="already-filled"]').val(), 'this is filled in');
+      assert.equal(view.$('[name="empty"]').val(), '');
       assert.equal(view.$('[name="has-data-novalue"]').val(), '');
       assert.equal(view.$('[name="name-only"]').val(), 'name only value');
       assert.equal(view.$('[name="name-preferred"]').val(), 'name preferred');
       assert.equal(view.$('[name="not-filled-but-saved"]').val(), '');
       assert.equal(view.$('[name="textarea"]').val(), 'the value for the text area');
+
+      // Prefilling a form element should not cause an `engaged` event.
+      assert.isFalse(view._engageFlowEventsForm.called);
     });
 
     it('formPrefill saves all input elements w/ name/id', () => {
@@ -71,9 +93,11 @@ define(function (require, exports, module) {
       view.$('[name="textarea"]').val('the value for the text area updated');
 
       view.destroy();
+
       assert.deepEqual(formPrefill.toJSON(), {
         'id-only': 'id only value updated',
         'already-filled': 'this is filled in', //eslint-disable-line
+        'empty': '',
         'has-data-novalue': 'has-data-novalue value',
         'name-only': 'name only value updated',
         'name-preferred': 'name preferred updated',